// src/lib/game-data/silk-data/crests-silk.generated.ts
export type CrestNameSilk =
	| 'Hunter'
	| 'Hunter_v2'
	| 'Hunter_v3'
	| 'Reaper'
	| 'Wanderer'
	| 'Warrior'
	| 'Witch'
	| 'Toolmaster'
	| 'Spell'
	| 'Cursed'
	| 'Cloakless';